%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.NonDocumentedObjects.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{Non-documented objects}
\alias{Non-documented objects}
\title{Non-documented objects}


% Other missing docs
\alias{IlluminaClusterPositionsTextFile}
\alias{readCentroidMatrix}
\alias{readCentroidMatrix.IlluminaClusterPositionsTextFile}
\alias{calculateExpectedRatiosByCentroids}
\alias{calculateExpectedRatiosByCentroids.matrix}
\alias{readUnitGenomePositions}
\alias{readUnitGenomePositions.IlluminaClusterPositionsTextFile}
\alias{calculateThetaRFromXY}
\alias{calculateThetaRFromXY.array}
\alias{calculateThetaRFromXY.matrix}
\alias{IlluminaIDatFile}
\alias{IlluminaIDatSet}
\alias{extractMatrix.IlluminaIDatFile}
\alias{extractMatrix.IlluminaIDatSet}
\alias{getPlatform.IlluminaIDatFile}
\alias{getPlatform.IlluminaIDatSet}
\alias{readQuants}
\alias{readQuants.IlluminaIDatFile}
\alias{readSignals}
\alias{readSignals.IlluminaIDatFile}
\alias{IlluminaGenomeStudioTextFile}
\alias{as.character.IlluminaGenomeStudioTextFile}
\alias{exportAromaUgpFile}
\alias{exportAromaUgpFile.IlluminaGenomeStudioTextFile}
\alias{exportTotalAndFracB}
\alias{exportTotalAndFracB.IlluminaGenomeStudioTextFile}
\alias{exportUnitNamesFile}
\alias{exportUnitNamesFile.IlluminaGenomeStudioTextFile}
\alias{extractTotalAndFracB}
\alias{extractTotalAndFracB.IlluminaGenomeStudioTextFile}
\alias{getChipType.IlluminaGenomeStudioTextFile}
\alias{getDataSetName}
\alias{getDataSetName.IlluminaGenomeStudioTextFile}
\alias{getDataTypes}
\alias{getDataTypes.IlluminaGenomeStudioTextFile}
\alias{getPlatform.IlluminaGenomeStudioTextFile}
\alias{getSampleNames.IlluminaGenomeStudioTextFile}
\alias{getUnitNames.IlluminaGenomeStudioTextFile}
\alias{getUnitNamesFile.IlluminaGenomeStudioTextFile}
\alias{nbrOfSamples}
\alias{nbrOfSamples.IlluminaGenomeStudioTextFile}
\alias{nbrOfUnits.IlluminaGenomeStudioTextFile}
\alias{readSampleData}
\alias{readSampleData.IlluminaGenomeStudioTextFile}
\alias{readUnitGenomePositions.IlluminaGenomeStudioTextFile}

\description{
  This page contains aliases for all "non-documented" objects that
  \code{R CMD check} detects in this package.

  Almost all of them are \emph{generic} functions that have specific
  document for the corresponding method coupled to a specific class.
  Other functions are re-defined by \code{setMethodS3()} to
  \emph{default} methods. Neither of these two classes are non-documented
  in reality.
  The rest are deprecated methods.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}


\keyword{documentation}
\keyword{internal}
